Fixed RNDISEthernet demos crashing when calculating checksums for Ethernet/TCP packets of more than ~500 bytes due to an overflow in the checksum calculation loop (thanks to Kevin Malec).
Removed string Attributes from the Service Discovery Protocol code to minimise the potential points of failure while the base code is being debugged.

+++ Demos/Device/LowLevel/RNDISEthernet/Lib/Ethernet.c
@@ -125,7 +125,7 @@ uint16_t Ethernet_Checksum16(void* Data, uint16_t Bytes)
 	uint16_t* Words    = (uint16_t*)Data;
 	uint32_t  Checksum = 0;
 
-	for (uint8_t CurrWord = 0; CurrWord < (Bytes >> 1); CurrWord++)
+	for (uint16_t CurrWord = 0; CurrWord < (Bytes >> 1); CurrWord++)
 	  Checksum += Words[CurrWord];
 	  
 	while (Checksum & 0xFFFF0000)
